Test: 멤버의 workspaceCount에 증가/감소에 대한 테스트 추가
- workspaceDelete에 대한 로직 개선 필요
- 기존 WorkspaceCreateServiceTest -> Deprecated
  - 해당 테스트는 다시 @SpringBootTest를 이용하는 테스트로 바꿀 예정
- workspaceCount에 대한 테스트는 새로운 WorkspaceCreateTest에 작성
