+++ src/test/java/com/uranus/taskmanager/api/workspace/service/WorkspaceCreateServiceTest.java
@@ -1,210 +1,148 @@
 package com.uranus.taskmanager.api.workspace.service;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.*;
 
-import java.util.Optional;
-
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.uranus.taskmanager.api.authentication.dto.LoginMember;
 import com.uranus.taskmanager.api.member.domain.Member;
-import com.uranus.taskmanager.api.member.repository.MemberRepository;
-import com.uranus.taskmanager.api.security.PasswordEncoder;
-import com.uranus.taskmanager.api.workspace.domain.Workspace;
+import com.uranus.taskmanager.api.member.exception.WorkspaceCreationLimitExceededException;
 import com.uranus.taskmanager.api.workspace.dto.request.WorkspaceCreateRequest;
+import com.uranus.taskmanager.api.workspace.dto.request.WorkspaceDeleteRequest;
 import com.uranus.taskmanager.api.workspace.dto.response.WorkspaceCreateResponse;
-import com.uranus.taskmanager.api.workspace.exception.WorkspaceCodeCollisionHandleException;
-import com.uranus.taskmanager.api.workspace.repository.WorkspaceRepository;
-import com.uranus.taskmanager.api.workspace.util.WorkspaceCodeGenerator;
-import com.uranus.taskmanager.api.workspacemember.WorkspaceRole;
-import com.uranus.taskmanager.api.workspacemember.domain.WorkspaceMember;
-import com.uranus.taskmanager.api.workspacemember.repository.WorkspaceMemberRepository;
-import com.uranus.taskmanager.fixture.dto.LoginMemberDtoFixture;
-import com.uranus.taskmanager.fixture.dto.WorkspaceCreateDtoFixture;
-import com.uranus.taskmanager.fixture.entity.MemberEntityFixture;
-import com.uranus.taskmanager.fixture.entity.WorkspaceEntityFixture;
-import com.uranus.taskmanager.fixture.entity.WorkspaceMemberEntityFixture;
+import com.uranus.taskmanager.helper.ServiceIntegrationTestHelper;
 
-@ExtendWith(MockitoExtension.class)
-class WorkspaceCreateServiceTest {
+class WorkspaceCreateServiceTest extends ServiceIntegrationTestHelper {
 
-	@InjectMocks
-	private CheckCodeDuplicationService workspaceCreateService;
-
-	@Mock
-	private WorkspaceCodeGenerator workspaceCodeGenerator;
-	@Mock
-	private WorkspaceRepository workspaceRepository;
-	@Mock
-	private MemberRepository memberRepository;
-	@Mock
-	private WorkspaceMemberRepository workspaceMemberRepository;
-	@Mock
-	private PasswordEncoder passwordEncoder;
-
-	WorkspaceEntityFixture workspaceEntityFixture;
-	MemberEntityFixture memberEntityFixture;
-	WorkspaceMemberEntityFixture workspaceMemberEntityFixture;
-	LoginMemberDtoFixture loginMemberDtoFixture;
-	WorkspaceCreateDtoFixture workspaceCreateDtoFixture;
-
-	@BeforeEach
-	public void setup() {
-		workspaceMemberEntityFixture = new WorkspaceMemberEntityFixture();
-		memberEntityFixture = new MemberEntityFixture();
-		loginMemberDtoFixture = new LoginMemberDtoFixture();
-		workspaceEntityFixture = new WorkspaceEntityFixture();
-		workspaceCreateDtoFixture = new WorkspaceCreateDtoFixture();
+	@AfterEach
+	public void tearDown() {
+		databaseCleaner.execute();
 	}
 
 	@Test
-	@DisplayName("워크스페이스 생성에는 생성 요청과 로그인 멤버를 필요로 한다")
-	void test1() {
+	@DisplayName("워크스페이스 생성 시 멤버의 워크스페이스 카운트가 증가한다")
+	void createWorkspace_increasesWorkspaceCount() {
 		// given
-		WorkspaceCreateRequest request = workspaceCreateDtoFixture.createWorkspaceCreateRequest(null);
-		String code = "TESTCODE";
+		Member member = memberRepository.save(
+			Member.builder()
+				.loginId("member1")
+				.email("member1@test.com")
+				.password("password1234!")
+				.build()
+		);
 
-		Workspace workspace = workspaceEntityFixture.createWorkspace("testcode");
-		Member member = memberEntityFixture.createMember("user123", "test@test.com");
-		LoginMember loginMember = loginMemberDtoFixture.createLoginMemberDto(1L, "user123", "test@test.com");
-
-		when(memberRepository.findById(loginMember.getId())).thenReturn(Optional.of(member));
-		when(workspaceCodeGenerator.generateWorkspaceCode()).thenReturn(code);
-		when(workspaceRepository.save(any(Workspace.class))).thenReturn(workspace);
+		WorkspaceCreateRequest request = WorkspaceCreateRequest.builder()
+			.name("workspace1")
+			.description("description1")
+			.build();
 
 		// when
-		WorkspaceCreateResponse response = workspaceCreateService.createWorkspace(request, loginMember.getId());
+		workspaceCreateService.createWorkspace(request, member.getId());
 
 		// then
-		assertThat(response).isNotNull();
-		verify(memberRepository, times(1)).findById(1L);
-	}
-
-	@DisplayName("워크스페이스 생성을 성공하면 WorkspaceResponse를 반환한다")
-	void test2() {
-		// given
-		WorkspaceCreateRequest request = workspaceCreateDtoFixture.createWorkspaceCreateRequest(null);
-		String code = "TESTCODE";
-
-		Workspace workspace = workspaceEntityFixture.createWorkspace("testcode");
-		Member member = memberEntityFixture.createMember("user123", "test@test.com");
-		LoginMember loginMember = loginMemberDtoFixture.createLoginMemberDto(1L, "user123", "test@test.com");
-
-		when(memberRepository.findById(loginMember.getId())).thenReturn(Optional.of(member));
-		when(workspaceCodeGenerator.generateWorkspaceCode()).thenReturn(code);
-		when(workspaceRepository.save(any(Workspace.class))).thenReturn(workspace);
-
-		// when
-		WorkspaceCreateResponse response = workspaceCreateService.createWorkspace(request, loginMember.getId());
-
-		// then
-		assertThat(response).isNotNull();
-		assertThat(response).isInstanceOf(WorkspaceCreateResponse.class);
-		verify(workspaceRepository, times(1)).save(any(Workspace.class));
+		Member updatedMember = memberRepository.findById(member.getId()).get();
+		assertThat(updatedMember.getWorkspaceCount()).isEqualTo(1);
 	}
 
 	@Test
-	@DisplayName("워크스페이스 코드가 중복될 때 최대 재시도 횟수(5회)를 소진하면 예외가 발생한다")
-	void test3() {
+	@DisplayName("워크스페이스 생성 시 멤버의 워크스페이스 카운트가 증가한다(워크스페이스 2개 생성)")
+	void createTwoWorkspaces_increasesWorkspaceCount() {
 		// given
-		WorkspaceCreateRequest request = workspaceCreateDtoFixture.createWorkspaceCreateRequest(null);
+		Member member = memberRepository.save(
+			Member.builder()
+				.loginId("member1")
+				.email("member1@test.com")
+				.password("password1234!")
+				.build()
+		);
 
-		Member member = memberEntityFixture.createMember("user123", "test@test.com");
-		LoginMember loginMember = loginMemberDtoFixture.createLoginMemberDto(1L, "user123", "test@test.com");
-		when(memberRepository.findById(loginMember.getId())).thenReturn(Optional.of(member));
+		WorkspaceCreateRequest request1 = WorkspaceCreateRequest.builder()
+			.name("workspace1")
+			.description("description1")
+			.build();
 
-		when(workspaceCodeGenerator.generateWorkspaceCode())
-			.thenReturn("WORK123", "WORK124", "WORK125", "WORK126", "WORK127");
-		when(workspaceRepository.existsByCode(anyString())).thenReturn(true);
+		WorkspaceCreateRequest request2 = WorkspaceCreateRequest.builder()
+			.name("workspace2")
+			.description("description2")
+			.build();
+
+		// when
+		workspaceCreateService.createWorkspace(request1, member.getId());
+		workspaceCreateService.createWorkspace(request2, member.getId());
+
+		// then
+		Member updatedMember = memberRepository.findById(member.getId()).get();
+		assertThat(updatedMember.getWorkspaceCount()).isEqualTo(2);
+	}
+
+	@Test
+	@DisplayName("워크스페이스 50개 생성 후 추가 생성 시 예외가 발생한다")
+	void createWorkspace_throwsException_whenLimitReached() {
+		// given
+		Member member = memberRepository.save(
+			Member.builder()
+				.loginId("member1")
+				.email("member1@test.com")
+				.password("password1")
+				.build()
+		);
+
+		// Create 50 workspaces
+		for (int i = 0; i < 50; i++) {
+			WorkspaceCreateRequest request = WorkspaceCreateRequest.builder()
+				.name("workspace" + i)
+				.description("description" + i)
+				.build();
+			workspaceCreateService.createWorkspace(request, member.getId());
+		}
 
 		// when & then
-		assertThatThrownBy(() -> workspaceCreateService.createWorkspace(request, loginMember.getId()))
-			.isInstanceOf(WorkspaceCodeCollisionHandleException.class)
-			.hasMessageContaining("Failed to solve workspace code collision");
+		WorkspaceCreateRequest request51 = WorkspaceCreateRequest.builder()
+			.name("workspace51")
+			.description("description51")
+			.build();
+
+		assertThatThrownBy(() -> workspaceCreateService.createWorkspace(request51, member.getId()))
+			.isInstanceOf(WorkspaceCreationLimitExceededException.class);
 	}
 
+	/*
+	 * Todo
+	 *  - 워크스페이스의 주인(생성자)의 로그인 Id를 통해서 주인 멤버 찾기
+	 *  - 해당 멤버의 workspaceCount 감소
+	 *  - 추후에 로직 개선 필요(DD를 적용하면 전체적으로 변할 듯)
+	 *  - 주인(Owner) Role 추가 -> 그냥 워크스페이스 삭제를 주인만 가능하도록 제한하면 더 쉬울 듯
+	 *  - -> 컨트롤러에서 LoginMember(id)를 읽어와서 사용하면 됨
+	 *  <br>
+	 *  - 해당 workspaceDelete의 로직을 변경하고 나서 테스트 실행
+	 */
+	@Disabled
 	@Test
-	@DisplayName("워크스페이스 생성 시 WorkspaceMember도 생성되고 저장된다")
-	void test4() {
+	@DisplayName("워크스페이스 삭제 시 멤버의 워크스페이스 카운트가 감소한다")
+	void deleteWorkspace_decreasesWorkspaceCount() {
 		// given
-		WorkspaceCreateRequest request = workspaceCreateDtoFixture.createWorkspaceCreateRequest(null);
-		String code = "TESTCODE";
+		Member member = memberRepository.save(
+			Member.builder()
+				.loginId("member1")
+				.email("member1@test.com")
+				.password("password1234!")
+				.build()
+		);
 
-		Workspace workspace = workspaceEntityFixture.createWorkspace("testcode");
-		Member member = memberEntityFixture.createMember("user123", "test@test.com");
-		LoginMember loginMember = loginMemberDtoFixture.createLoginMemberDto(1L, "user123", "test@test.com");
-		WorkspaceMember workspaceMember = workspaceMemberEntityFixture.createAdminWorkspaceMember(member, workspace);
-		when(memberRepository.findById(loginMember.getId())).thenReturn(Optional.of(member));
-		when(workspaceCodeGenerator.generateWorkspaceCode()).thenReturn(code);
-		when(workspaceRepository.save(any(Workspace.class))).thenReturn(workspace);
-		when(workspaceMemberRepository.save(any(WorkspaceMember.class))).thenReturn(workspaceMember);
+		WorkspaceCreateRequest request = WorkspaceCreateRequest.builder()
+			.name("workspace1")
+			.description("description1")
+			.build();
+
+		WorkspaceCreateResponse response = workspaceCreateService.createWorkspace(request, member.getId());
 
 		// when
-		workspaceCreateService.createWorkspace(request, loginMember.getId());
+		workspaceCommandService.deleteWorkspace(new WorkspaceDeleteRequest(), response.getCode());
 
 		// then
-		verify(workspaceMemberRepository, times(1)).save(any(WorkspaceMember.class));
-	}
-
-	@Test
-	@DisplayName("워크스페이스 생성 시 WorkspaceMember의 별칭은 생성자의 이메일로 설정된다")
-	void test5() {
-		// given
-		WorkspaceCreateRequest request = workspaceCreateDtoFixture.createWorkspaceCreateRequest(null);
-		String code = "TESTCODE";
-
-		Workspace workspace = workspaceEntityFixture.createWorkspace("testcode");
-		Member member = memberEntityFixture.createMember("user123", "test@test.com");
-		LoginMember loginMember = loginMemberDtoFixture.createLoginMemberDto(1L, "user123", "test@test.com");
-		WorkspaceMember adminWorkspaceMember = workspaceMemberEntityFixture.createAdminWorkspaceMember(member,
-			workspace);
-
-		when(memberRepository.findById(loginMember.getId())).thenReturn(Optional.of(member));
-		when(workspaceCodeGenerator.generateWorkspaceCode()).thenReturn(code);
-		when(workspaceRepository.save(any(Workspace.class))).thenReturn(workspace);
-		when(workspaceMemberRepository.save(any(WorkspaceMember.class))).thenReturn(adminWorkspaceMember);
-
-		// when
-		workspaceCreateService.createWorkspace(request, loginMember.getId());
-
-		// then
-		verify(workspaceMemberRepository, times(1)).save(argThat(workspaceMember ->
-			workspaceMember.getNickname().equals("test@test.com")
-		));
-	}
-
-	@Test
-	@DisplayName("워크스페이스 생성 시 WorkspaceMember의 권한은 ADMIN으로 설정된다")
-	void test6() {
-		// given
-		WorkspaceCreateRequest request = workspaceCreateDtoFixture.createWorkspaceCreateRequest(null);
-		String code = "TESTCODE";
-
-		Workspace workspace = workspaceEntityFixture.createWorkspace("testcode");
-		Member member = memberEntityFixture.createMember("user123", "test@test.com");
-		LoginMember loginMember = loginMemberDtoFixture.createLoginMemberDto(1L, "user123", "test@test.com");
-		WorkspaceMember adminWorkspaceMember = workspaceMemberEntityFixture.createAdminWorkspaceMember(member,
-			workspace);
-
-		when(memberRepository.findById(loginMember.getId())).thenReturn(Optional.of(member));
-		when(workspaceCodeGenerator.generateWorkspaceCode()).thenReturn(code);
-		when(workspaceRepository.save(any(Workspace.class))).thenReturn(workspace);
-		when(workspaceMemberRepository.save(any(WorkspaceMember.class))).thenReturn(adminWorkspaceMember);
-
-		// when
-		workspaceCreateService.createWorkspace(request, loginMember.getId());
-
-		// then
-		verify(workspaceMemberRepository, times(1)).save(argThat(workspaceMember ->
-			workspaceMember.getRole().equals(WorkspaceRole.ADMIN)
-		));
+		Member updatedMember = memberRepository.findById(member.getId()).get();
+		assertThat(updatedMember.getWorkspaceCount()).isEqualTo(0);
 	}
 }
